# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flight-review-server",
    "description": "Flight Review v2 — HTTP API server",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "operationId": "get_stats",
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "description": "Column to group by (validated against allowlist in the handler).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Time period: \"7d\", \"30d\", \"90d\", \"1y\", \"all\".",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of groups to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "vehicle_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ver_hw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ver_sw_release_str",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vibration_status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated log statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid aggregation request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "StatRow": {
        "type": "object",
        "required": [
          "group",
          "count"
        ],
        "properties": {
          "avg_flight_duration_s": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "avg_max_speed": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "group": {
            "type": "string"
          },
          "total_flight_hours": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "StatsResponse": {
        "type": "object",
        "required": [
          "group_by",
          "period",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatRow"
            }
          },
          "group_by": {
            "type": "string"
          },
          "period": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}